[refactor] Adjust favorability weight
The favorability weight in the analysis module has been adjusted from 0.2 to 0.1. This change will affect the data parsing and analysis of candidate favorability ratings.

The reasoning behind this is as we get closer to the election date the distinction between the likability of a candidate and likelihood of voting for a candidate become confused.