Update date parsing and favorability weight
Updated the way dates are parsed in the 'created_at' column of the DataFrame to use a specified format instead of default parsing. This change provides more control over how dates are interpreted, especially when dealing with different date formats.

Also, adjusted the FAVORABILITY_WEIGHT from 0.2 to 0.25 in config.py for better analysis results. This tweak allows us to give slightly more importance to favorability in our combined analysis.