[refactor] Simplify color scale creation and comment out grouped analysis
Refactored the way color scales are created in line charts. Instead of manually defining the range, a dictionary is now used to map column names to their respective colors. This makes the code more maintainable and easier to understand.

Additionally, the section for creating a grouped analysis chart has been commented out. This change does not affect functionality as it seems this part was not being used.