refactor: Simplify period definition and expand options
The periods for analysis in the `analysis.py` file have been refactored to be dynamically generated from a list of strings in the `config.py` file. This change simplifies the code and makes it easier to add or remove periods for analysis. The list of periods has also been expanded with additional options.