[refactor] Change favorability calculation method
Refactored the way favorability is calculated in the analysis. Instead of calculating a differential between favorable and unfavorable weights, we now calculate favorability as a percentage of total weight. This change also affects how combined scores are calculated in the combine_analysis function, where we now use these new favorability scores instead of differentials.